requirements.txt
streamlit
ultralytics
opencv-python
pillow
numpy
torch
torchvision